sending throughput from monitoring instance
