+++ src/ServiceControl.Monitoring/HostApplicationBuilderExtensions.cs
@@ -16,6 +16,7 @@ using NLog.Extensions.Logging;
 using NServiceBus;
 using NServiceBus.Configuration.AdvancedExtensibility;
 using NServiceBus.Features;
+using NServiceBus.Metrics;
 using NServiceBus.Transport;
 using QueueLength;
 using ServiceControl.Monitoring.Infrastructure.BackgroundTasks;
@@ -63,13 +64,13 @@ public static class HostApplicationBuilderExtensions
 
         services.AddLicenseCheck();
 
-        ConfigureEndpoint(endpointConfiguration, onCriticalError, transportCustomization, settings);
+        ConfigureEndpoint(endpointConfiguration, onCriticalError, transportCustomization, settings, services);
         hostBuilder.UseNServiceBus(endpointConfiguration);
 
         hostBuilder.AddAsyncTimer();
     }
 
-    static void ConfigureEndpoint(EndpointConfiguration config, Func<ICriticalErrorContext, CancellationToken, Task> onCriticalError, ITransportCustomization transportCustomization, Settings settings)
+    static void ConfigureEndpoint(EndpointConfiguration config, Func<ICriticalErrorContext, CancellationToken, Task> onCriticalError, ITransportCustomization transportCustomization, Settings settings, IServiceCollection services)
     {
         if (!string.IsNullOrWhiteSpace(settings.LicenseFileText))
         {
@@ -85,6 +86,22 @@ public static class HostApplicationBuilderExtensions
         };
 
         transportCustomization.CustomizeMonitoringEndpoint(config, transportSettings);
+
+        var serviceControlThroughputDataQueue = settings.ServiceControlThroughputDataQueue;
+        if (!string.IsNullOrWhiteSpace(serviceControlThroughputDataQueue))
+        {
+            if (serviceControlThroughputDataQueue.IndexOf("@") >= 0)
+            {
+                serviceControlThroughputDataQueue = serviceControlThroughputDataQueue.Substring(0, serviceControlThroughputDataQueue.IndexOf("@"));
+            }
+
+            var routing = new RoutingSettings(config.GetSettings());
+            routing.RouteToEndpoint(typeof(RecordEndpointThroughputData), serviceControlThroughputDataQueue);
+
+            services.AddSingleton<ReportThroughputFeatureStartup>();
+            config.EnableFeature<ReportThroughputFeature>();
+        }
+
 
         if (settings.EnableInstallers)
         {
