+++ src/ServiceControl.Monitoring/Infrastructure/WebApi/HostApplicationBuilderExtensions.cs
@@ -4,11 +4,14 @@ using System.Reflection;
 using Infrastructure;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
+using ServiceControl.Monitoring.Infrastructure.Api;
 
 public static class HostApplicationBuilderExtensions
 {
     public static void AddServiceControlMonitoringApi(this IHostApplicationBuilder hostBuilder)
     {
+        hostBuilder.Services.AddSingleton<IEndpointMetricsApi, EndpointMetricsApi>();
+
         var controllers = hostBuilder.Services.AddControllers(options =>
         {
             options.Filters.Add<XParticularVersionHttpHandler>();
